Add files via upload
Beta version of TrackMPD with resuspension based on FVCOM application. Replace the old files in the previous TRackMPD with these new matlab programs and you are good to go.

In this version:
(1) Users can adopt Kv and Kh as variables (time- and node-dependent) instead of constant;
(2) Reproduce resuspension of a single particle based on the concurrent bottom stress when it deposited.